Implement as-to-kafka proof-of-concept
as-to-kafka is a proof-of-concept app that continuously reads
new data from an Aerospike cluster and publishes it to a Kafka topic.
The records and keys are encoding using msgpack.

Usage: ./as-to-kafka <broker> <topic> <hosts> <ns> <set>

Where:
  broker - comma-separated list of Kafka bootstrap brokers (host or host:port)
  topic  - Kafka topic to publish to
  hosts  - comma-separated list of Aerospike cluster nodes (host or host:port)
  ns     - Aerospike namespace to stream data from
  set    - Set name within the Aerospike namespace

consumer is a simple demo app that consumes a Kafka topic and prints the
data out to the console as decoded msgpack object or hexdump.

Usage: ./consumer [options] <topic[:part]> <topic[:part]>..